added CFAUPresetRef, CFAUPresetCreate(), and CFAUPresetRelease()
There was a fundamental problem with my previous approach to the
CFArrayCallbacks for AUPresets in that they didn't use the
CoreFoundation memory model.  Retain would actually allocate a new
instance, and Release would always deallocate the instance, so
multiple retains would lead to memory leaks and multiple releases
could lead to crashes or other memory misbehavior.  So to solve this
problem, I created a new object called CFAUPreset that contains an
AUPreset struct, and then the functions CFAUPresetCreate() and
CFAUPresetRelease() implement a more CoreFoundation handling of their
memory and lifespan.  The CFArray retain and release callbacks have
been updated accordingly to make proper use of these changes.