improved the debug reporting done in DfxSettings::correctEndian()
also, this was missing from the notes for the previous check-in,
but is important:  I fixed a bug in DfxSettings::correctEndian()
where, instead of iterating through each parameter assignment
structure and byte-reordering each value, it would instead on each
iteration (for each parameter) byte-reorder only the first structure,
over and over and over.  So in the end, the first structure may or may
not be reordered (depending whether the plugin has an odd or even
number of parameters), and everything else would definitely not be
reordered.  I've fixed this now, but it unfortunately is going to
affect backwards compatibility.  Not for settings saved on PPC
machines, but for all other cases, the settings data will be corrupt
in the parameter assignments section.  I'm deciding to not worry about
this since the parameter MIDI assignments aren't the most crucial
piece in the settings data, and while annoying, I don't think it's a
huge problem for them to have been corrupt, so I'm not going to do
anything about recognizing and correcting old settings.