Add workaround for vstgui not compiling in mingw32 any more. I think this is really a problem with mingw; see this thread: https://sourceforge.net/p/mingw-w64/mailman/mingw-w64-public/thread/20200920130036.1e6b6508%40ext.localdomain/
The workaround is certainly not correct (it will also apply to the headers we include!), but since we compile in C++ mode and don't reference the C++ overloads (at different types), it works ok.
Should revisit this to see if the define is still needed in the next mingw release.