Replace MT19937 random engine with PCG, which is a lot faster and (especially) smaller. Supposedly PCG has better spectral properties (passes more statistical tests) than MT too, but I think the main point is just to be lightweight.
My benchmark results:

MT:
1000000000 iters in 14.7932 sec
sizeof rg: 2512

PCG:
1000000000 iters in 1.5210 sec
sizeof rg: 16

The results are less dramatic if you're generating 64-bit integers, probably because the underlying engine has 32-bit output, but it was still better on all the cases I tried (and does work when constructed with <uint64_t>(0, ~0)).
There are versions of this with 64-bit output we could try if we're interested in fine tuning, but it probably doesn't matter?